refactor cheats
